Se agrego funcion para subir las imagenes y obtener la URL de mineral

+++ routes/mineral.js
@@ -1,7 +1,12 @@
 const express = require('express');
 const router = express.Router();
 const db = require('../config/bd');
+const createUpload = require('../middlewares/upload');
+const drive = require('../config/drive');
+const fs = require('fs');
+const upload = createUpload('ID_MINERAL');
 
+const id_Carpeta_Drive = '1a0dUrckJ94wYICyNS_tvj3-DUcf75O9f';
 
 //Obtener todos los minerales
 router.get('/', async (req, res) => {
@@ -28,11 +33,35 @@ router.get('/:ID_MINERAL', async (req, res) => {
 });
 
 //Crear un nuevo mineral
-router.post('/', async (req, res) => {
+router.post('/', upload.single('FILE'), async (req, res) => {
     const { ID_MINERAL, N_BARRANTES, COLECCION, NOMBRE_MINERAL, CANTIDAD, GRUPO_MINERALOGICO, REGION, SUBGRUPO, COMPOSICION,
-        CARACTERISTICAS, COLECTOR, OBSERVACIONES, UBICACION, FOTO } = req.body;
+        CARACTERISTICAS, COLECTOR, OBSERVACIONES, UBICACION } = req.body;
+
+    var FOTO = '';
 
     try {
+        //Subir imagen a Google Drive
+        if (req.file) {
+            const filePath = req.file.path;
+
+            const response = await drive.files.create({
+                requestBody: {
+                    name: req.file.filename,
+                    mimeType: req.file.mimetype,
+                    parents: [id_Carpeta_Drive], //ID de la carpeta de Google Drive
+                },
+                media: {
+                    mimeType: req.file.mimetype,
+                    body: fs.createReadStream(filePath),
+                },
+            });
+
+            //Eliminar archivo temporal del servidor
+            fs.unlinkSync(filePath);
+
+            FOTO = `https://lh3.googleusercontent.com/d/${response.data.id}`;
+        }
+
         const [result] = await db.query(
             'INSERT INTO mineral (ID_MINERAL, N_BARRANTES, COLECCION, NOMBRE_MINERAL, CANTIDAD, GRUPO_MINERALOGICO, REGION, SUBGRUPO, COMPOSICION, CARACTERISTICAS, COLECTOR, OBSERVACIONES, UBICACION, FOTO) VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,?)',
             [ID_MINERAL, N_BARRANTES, COLECCION, NOMBRE_MINERAL, CANTIDAD, GRUPO_MINERALOGICO, REGION, SUBGRUPO, COMPOSICION,
@@ -49,12 +78,50 @@ router.post('/', async (req, res) => {
 });
 
 //Actulizar los datos del mineral
-router.put('/:ID_MINERALPARAM', async (req, res) => {
+router.put('/:ID_MINERALPARAM', upload.single('FILE'), async (req, res) => {
     const { ID_MINERALPARAM } = req.params;
     const { ID_MINERAL, N_BARRANTES, COLECCION, NOMBRE_MINERAL, CANTIDAD, GRUPO_MINERALOGICO, REGION, SUBGRUPO, COMPOSICION,
-        CARACTERISTICAS, COLECTOR, OBSERVACIONES, UBICACION, FOTO } = req.body;
+        CARACTERISTICAS, COLECTOR, OBSERVACIONES, UBICACION } = req.body;
 
+    var FOTO = '';
     try {
+        //Obtener la imagen actual desde la BD
+        const [rows] = await db.query('SELECT FOTO FROM mineral WHERE ID_MINERAL = ?', [ID_MINERAL]);
+        if (rows.length === 0) {
+            return res.status(404).json({ error: `El ID del mineral: ${ID_MINERAL} no fue encontrado` });
+        }
+
+        const currentFotoUrl = rows[0].FOTO;
+        const currentFileId = currentFotoUrl.split('/d/')[1];
+
+        //Subir imagen a Google Drive si se proporciona un archivo
+        if (req.file) {
+            const filePath = req.file.path;
+
+            if (currentFileId) {
+                await drive.files.delete({ fileId: currentFileId }).catch(console.error); //Si se reemplaza la imagen, borra la que tiene vinculada para agregar la nueva
+            }
+
+            const response = await drive.files.create({
+                requestBody: {
+                    name: req.file.filename,
+                    mimeType: req.file.mimetype,
+                    parents: [id_Carpeta_Drive], //ID de la carpeta de Google Drive
+                },
+                media: {
+                    mimeType: req.file.mimetype,
+                    body: fs.createReadStream(filePath),
+                },
+            });
+
+            //Eliminar archivo temporal del servidor
+            fs.unlinkSync(filePath);
+
+            FOTO = `https://lh3.googleusercontent.com/d/${response.data.id}`;
+        } else {
+            FOTO = currentFotoUrl; //Mantener la imagen actual si no se proporciona una nueva
+        }
+
         const [result] = await db.query(
             'UPDATE mineral SET ID_MINERAL = ?, N_BARRANTES = ?, COLECCION = ?, NOMBRE_MINERAL = ?, CANTIDAD = ?, GRUPO_MINERALOGICO = ?, REGION = ?, SUBGRUPO = ?, COMPOSICION = ?, CARACTERISTICAS = ?, COLECTOR = ?, OBSERVACIONES = ?, UBICACION = ?, FOTO = ? WHERE ID_MINERAL = ?',
             [ID_MINERAL, N_BARRANTES, COLECCION, NOMBRE_MINERAL, CANTIDAD, GRUPO_MINERALOGICO, REGION, SUBGRUPO, COMPOSICION,
@@ -74,6 +141,21 @@ router.delete('/:ID_MINERAL', async (req, res) => {
     const { ID_MINERAL } = req.params;
 
     try {
+        //Obtener la imagen actual desde la BD
+        const [rows] = await db.query('SELECT FOTO FROM mineral WHERE ID_MINERAL = ?', [ID_MINERAL]);
+        if (rows.length === 0) {
+            return res.status(404).json({ error: `El ID del mineral: ${ID_MINERAL} no fue encontrado` });
+        }
+
+        const currentFotoUrl = rows[0].FOTO;
+        const currentFileId = currentFotoUrl.split('/d/')[1];
+
+        // Eliminar la imagen de Google Drive
+        if (currentFileId) {
+            await drive.files.delete({ fileId: currentFileId }).catch(console.error);
+        }
+
+        //Consulta para borrar los demas datos
         const [result] = await db.query('DELETE FROM mineral WHERE ID_MINERAL = ?', [ID_MINERAL]);
         if (result.affectedRows === 0) {
             return res.status(404).json({ error: `El ID del mineral: ${ID_MINERAL} no fue encontrado` });
@@ -81,6 +163,40 @@ router.delete('/:ID_MINERAL', async (req, res) => {
         res.status(200).json({ message: `El mineral fue eliminado correctamente` });
     } catch (error) {
         res.status(500).json({ error: 'Error al eliminar los datos del mineral' });
+    }
+});
+
+
+//Eliminar imagen de un mineral
+router.delete('/:ID_MINERAL/image', async (req, res) => {
+    const { ID_MINERAL } = req.params;
+
+    try {
+        // Obtener la URL de la imagen actual de la pieza desde la base de datos
+        const [rows] = await db.query('SELECT FOTO FROM mineral WHERE ID_MINERAL = ?', [ID_MINERAL]);
+        if (rows.length === 0) {
+            return res.status(404).json({ error: `El ID del mineral: ${ID_MINERAL} no fue encontrado` });
+        }
+        const currentFotoUrl = rows[0].FOTO;
+
+        if (!currentFotoUrl) {
+            return res.status(400).json({ error: 'El mineral no tiene una imagen asociada' });
+        }
+
+        // Extraer el ID del archivo de Google Drive desde la URL
+        const currentFileId = currentFotoUrl.split('/d/')[1];
+
+        // Eliminar la imagen de Google Drive
+        if (currentFileId) {
+            await drive.files.delete({ fileId: currentFileId }).catch(console.error);
+        }
+
+        // Actualizar el campo FOTO a NULL en la base de datos
+        await db.query('UPDATE mineral SET FOTO = "" WHERE ID_MINERAL = ?', [ID_MINERAL]);
+
+        res.status(200).json({ message: `La imagen del mineral con ID ${ID_MINERAL} fue eliminada correctamente` });
+    } catch (error) {
+        res.status(500).json({ error: 'Error al eliminar la imagen del mineral' });
     }
 });
